requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.19
SQLAlchemy==0.9.4
Tempita==0.5.2
Werkzeug==0.9.4
aniso8601==0.82
decorator==3.4.0
itsdangerous==0.24
pbr==0.8.0
pytz==2014.2
six==1.6.1
sqlalchemy-migrate==0.9
wsgiref==0.1.2
